requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
Pillow==7.0.0
pytesseract==0.3.1
soupsieve==1.9.5
Flask==0.12
requests==2.22.0
Flask-Cors==3.0.8